+++ integration-tests/xcm-test/src/ctx/e2e/xcm.ts
@@ -23,8 +23,9 @@ import {
 } from './xcm.utils';
 import { SetupCtx } from './types';
 
-import { getRouteKey } from '../utils';
+import { getRouteKey } from '../../utils/route';
 
+const SNAPSHOT_NAME = '__snapshots__/e2e.spec.ts';
 const TRANSFER_AMOUNT = '10';
 
 export const runXcm = (
@@ -38,9 +39,12 @@ export const runXcm = (
     networks: SetupCtx[];
     wallet: Wallet;
   }>,
-  options: { skip?: boolean } = {}
+  options: { skip?: boolean; sync?: boolean; snapshot?: boolean } = {}
 ) => {
   const itfn = options.skip ? it.skip : it;
+  const shouldSync = options.sync || false;
+  const shouldSnapshot = options.snapshot || false;
+
   itfn(
     name,
     async () => {
@@ -68,6 +72,7 @@ export const runXcm = (
 
       let extrinsic;
       if (calldata.type === CallType.Evm && chain.key === 'moonbeam') {
+        return;
         // Skipped for now
         extrinsic = moonbeam.toTransferExtrinsic(
           srcNetwork.api,
@@ -98,7 +103,9 @@ export const runXcm = (
       await destNetwork.chain.newBlock();
       const destEvents = await destNetwork.api.query.system.events();
       expect(checkIfProcessed(destEvents)).toBeTruthy();
-      expect([key, calldata.data]).toMatchSnapshot();
+
+      shouldSnapshot &&
+        expect([key, calldata.data]).toMatchSnapshot(SNAPSHOT_NAME);
 
       const postTransfer = await getTransfer(
         wallet,
@@ -113,14 +120,15 @@ export const runXcm = (
         postTransfer
       );
 
-      report.set(key, {
-        updated: Date.now(),
-        destination: {
-          fee: destinationFee.delta,
-          feeAsset: destinationFee.asset,
-          feeNative: destinationFee.deltaBn,
-        },
-      });
+      shouldSync &&
+        report.set(key, {
+          updated: Date.now(),
+          destination: {
+            fee: destinationFee.delta,
+            feeAsset: destinationFee.asset,
+            feeNative: destinationFee.deltaBn,
+          },
+        });
 
       console.table(
         [