+++ integration-tests/xcm-test/src/utils/route.ts
@@ -1,7 +1,5 @@
 import { AnyChain, AssetRoute } from '@galacticcouncil/xcm-core';
 
-import { writeFileSync } from 'fs';
-
 export const getRouteKey = (chain: AnyChain, route: AssetRoute) => {
   const { source, destination } = route;
   return [chain.key, destination.chain.key, source.asset.key].join('-');
@@ -19,19 +17,4 @@ export const getRouteInfo = (chain: AnyChain, route: AssetRoute) => {
     transferAsset.originSymbol,
     `[${transferAsset.key}]`,
   ].join(' ');
-};
-
-export const jsonFormatter = (_: any, nestedValue: any) => {
-  return typeof nestedValue === 'bigint' ? nestedValue.toString() : nestedValue;
-};
-
-export const findNestedKey = (obj: any, keyToFind: any) => {
-  const foundObj: any[] = [];
-  JSON.stringify(obj, (_, nestedValue) => {
-    if (nestedValue && nestedValue[keyToFind]) {
-      foundObj.push(nestedValue);
-    }
-    return nestedValue;
-  });
-  return foundObj[0];
 };