+++ packages/xcm-cfg/src/builders/BalanceBuilder.ts
@@ -98,13 +98,7 @@ function foreignAssets() {
       build: ({ address, asset, chain }) => {
         const ctx = chain as Parachain;
 
-        // TODO - use v4 locations
         const assetLocation = ctx.getAssetXcmLocation(asset);
-        const normalizeAssetLocation = normalizeInterior(
-          assetLocation!,
-          XcmVersion.v4
-        );
-
         if (!assetLocation) {
           throw new Error('Missing asset xcm location for ' + asset.key);
         }
@@ -112,7 +106,7 @@ function foreignAssets() {
         return new SubstrateQueryConfig({
           module: 'foreignAssets',
           func: 'account',
-          args: [normalizeAssetLocation, address],
+          args: [assetLocation, address],
           transform: async (
             response: Option<PalletAssetsAssetAccount>
           ): Promise<bigint> => response.unwrapOrDefault().balance.toBigInt(),