+++ packages/xcm-cfg/src/configs/polkadot/hydration/index.ts
@@ -46,6 +46,7 @@ import {
 } from '../../../assets';
 import {
   acala,
+  acala_evm,
   ajuna,
   assetHub,
   astar,
@@ -67,6 +68,7 @@ import {
   zeitgeist,
 } from '../../../chains';
 import {
+  ContractBuilder,
   ExtrinsicBuilder,
   ExtrinsicBuilderV4,
   XcmTransferType,
@@ -79,6 +81,8 @@ import {
   toEthereumViaWormholeTemplate,
   toMoonbeamErc20Template,
   toZeitgeistErc20Template,
+  MRL_EXECUTION_FEE,
+  MRL_XCM_FEE,
 } from './templates';
 
 const toAcala: AssetRoute[] = [
@@ -482,7 +486,7 @@ const toMoonbeam: AssetRoute[] = [
       chain: moonbeam,
       asset: hdx,
       fee: {
-        amount: 0.835,
+        amount: 5,
         asset: hdx,
       },
     },
@@ -817,6 +821,53 @@ const toAjuna: AssetRoute[] = [
   }),
 ];
 
+const toAcalaViaWormhole: AssetRoute[] = [
+  new AssetRoute({
+    source: {
+      asset: dai_mwh,
+      balance: balance(),
+      fee: fee(),
+      destinationFee: {
+        balance: balance(),
+      },
+    },
+    destination: {
+      chain: acala_evm,
+      asset: dai_awh,
+      fee: {
+        amount: 0.06,
+        asset: aca,
+      },
+    },
+    extrinsic: ExtrinsicBuilder()
+      .utility()
+      .batchAll([
+        ExtrinsicBuilder().xTokens().transferMultiCurrencies(),
+        ExtrinsicBuilder().polkadotXcm().send().transact({
+          fee: MRL_EXECUTION_FEE,
+        }),
+      ]),
+    transact: {
+      chain: moonbeam,
+      fee: {
+        amount: MRL_XCM_FEE,
+        asset: glmr,
+        balance: balance(),
+      },
+      extrinsic: ExtrinsicBuilder()
+        .ethereumXcm()
+        .transact(
+          ContractBuilder()
+            .Batch()
+            .batchAll([
+              ContractBuilder().Erc20().approve(),
+              ContractBuilder().Wormhole().TokenBridge().transferTokens(),
+            ])
+        ),
+    },
+  }),
+];
+
 const toEthereumViaWormhole: AssetRoute[] = [
   toEthereumViaWormholeTemplate(dai_mwh, dai),
   toEthereumViaWormholeTemplate(weth_mwh, eth),
@@ -834,6 +885,7 @@ export const hydrationConfig = new ChainRoutes({
   chain: hydration,
   routes: [
     ...toAcala,
+    ...toAcalaViaWormhole,
     ...toAjuna,
     ...toAssetHub,
     ...toAstar,