+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @galacticcouncil/sdk
 
+## 5.0.1
+
+### Patch Changes
+
+- Fix asset location in router api
+
+## 5.0.0
+
+### Major Changes
+
+- Removing externalId & origin from Asset, using location instead (breaking change)
+
 ## 4.6.1
 
 ### Patch Changes